feat: use proper query to get my requests
